refactor(config): isolate manifest request handlers
